git-svn-id: svn://mmweb.animal.net.cn/nb_icesheet/trunk@21 ba629b7a-d26c-4841-822e-5ac455a76d44

+++ src/Definitions/CoodLocation.cpp
@@ -7,7 +7,7 @@
 
 #include "CoodLocation.h"
 
-CoodLocation::CoodLocation(unsigned p_x, unsigned p_y) {
+CoodLocation::CoodLocation(short unsigned p_x, short unsigned p_y) {
    x = p_x;
    y = p_y;
 
@@ -17,9 +17,9 @@ CoodLocation::~CoodLocation() {
 
 }
 
-unsigned CoodLocation::getX(){
+unsigned short CoodLocation::getX(){
     return x;
 }
-unsigned CoodLocation::getY(){
+unsigned short CoodLocation::getY(){
     return y;
 }
